refs #1 Added to support GSN's file. (agml extension file)
- This plugin can be used for diagrams on Astah Professional and Astah
GSN.
- astah-gsn.jar is added into astah.zip.
- An appropriate jar file will be used with reference to a file
extension.
